Integer types cleanup
Try to be a bit more strict about integer types, use off_t or int64_t for
file size and file offsets.

Replace safe_ato*() by safe_atoi32() and safe_atoi64(), fix integer types
at call sites to match.

+++ src/transcode.c
@@ -135,7 +135,7 @@ add_le32(uint8_t *dst, uint32_t val)
 }
 
 static void
-make_wav_header(struct transcode_ctx *ctx, size_t *est_size)
+make_wav_header(struct transcode_ctx *ctx, off_t *est_size)
 {
   uint32_t samplerate;
   uint32_t byte_rate;
@@ -310,7 +310,7 @@ transcode(struct transcode_ctx *ctx, struct evbuffer *evbuf, int wanted)
 }
 
 struct transcode_ctx *
-transcode_setup(struct media_file_info *mfi, size_t *est_size)
+transcode_setup(struct media_file_info *mfi, off_t *est_size)
 {
   struct transcode_ctx *ctx;
   int hdr_len;
